// role/src/interfaces/message-enums/message-enum.ts
export enum MessageEnum {
    CREATED = 'Role has been successfully created',
    ROLE_SEARCH_OK = 'Ok',
    ROLE_NOT_FOUND = 'No roles with this value',
    CONFLICT = 'Role with this value already exist',
    RELATION_NOT_FOUND = 'User with this id don\'t have any role',
    RELATION_CREATED = 'Role was successfully assigned to user',
    PRECONDITION_FAILED = 'Precondition failed',

}